Show combat mouse pointer

+++ actors.cc
@@ -1362,8 +1362,9 @@ Monster_actor *Monster_info::create
 	if (!equip_offset || equip_offset - 1 >= equip_cnt)
 		return (monster);	// Out of range.
 	Equip_record& rec = equip[equip_offset - 1];
-	for (int i = 0; i < sizeof(equip->elements)/sizeof(equip->elements[0]);
-								i++)
+	for (size_t i = 0;
+			i < sizeof(equip->elements)/sizeof(equip->elements[0]);
+			i++)
 		{			// Give equipment.
 		Equip_element& elem = rec.elements[i];
 		if (!elem.shapenum || 1 + rand()%100 > elem.probability)
